cli/requirements.txt
rich==13.7.0
playwright==1.40.0
python-dotenv==1.0.0
pyperclip==1.8.2
requests==2.31.0
pyautogui==0.9.54
google-generativeai==0.3.2

